requirements.txt
# Core LangChain and LangGraph dependencies
langchain>=0.1.0
langchain-core>=0.1.0
langchain-google-genai>=1.0.0
langgraph>=0.0.30

# Web framework
streamlit>=1.28.0

# Email service
sendgrid>=6.10.0

# API services
google-search-results>=2.4.2

# Environment and configuration
python-dotenv>=1.0.0

# Data validation
pydantic>=2.0.0,<3.0.0

# Python standard library extensions
typing-extensions>=4.5.0

langchain-openai